Move ApiGatewayApplication to package `com.matevskial.systemdesignplayground`
`com.matevskial.systemdesignplayground` represents the root package of this project. It is useful that the `main` classes for spring boot applications reside in the root package, so the spring application is able to scan the spring beans and configurations from other modules. For this to work without additional configuration, classes in the modules need to reside in a package with prefix `com.matevskial.systemdesignplayground`